requirements.txt
apache-airflow==2.8.4
black==24.10.0
confluent-kafka==2.6.1
dbt-snowflake==1.8.3
faker==30.8.2
flake8==7.1.1
pandas==2.2.3
plotly==5.24.1
psycopg2-binary==2.9.10
pyspark==3.5.2
python-dotenv==1.0.1
requests==2.32.3
snowflake-connector-python[pandas]==3.12.2
streamlit==1.40.2
